% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_message.R
\name{build_message}
\alias{build_message}
\title{Builds message to display for student}
\usage{
build_message(submission)
}
\arguments{
\item{submission}{A magrader submission object}
}
\value{
A character string suitable for using in the tutor exercise checker
  function message field.
}
\description{
Builds message to display for student
}
