% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_calls.R
\name{code_calls}
\alias{code_calls}
\title{Check whether code calls a desired function or object}
\usage{
code_calls(code, call, message = paste("I expected to see a call to", call))
}
\arguments{
\item{code}{The code to check, as a character string, or the output of a
previous check function. If you are using `code_calls()` within a tutor
document, set code to the special name `.answer`.}

\item{call}{The name of a function or object to check for as a character
string.}

\item{message}{A character string of feedback to display if the code fails
the check}
}
\value{
If the code passes the check, `code_calls()` returns the original
  code. If the code fails the check, `code_calls()` returns a failure object
  that contains a feedback message.
}
\description{
`code_calls()` is designed to be used in a magrittr pipe. It will first check
whether the code input has a;lready failed a previous check. If it has not,
`code_calls()` will check to see whether the code calls the given object or
function.
}
\examples{
\dontrun{
"head(mtcars)" \%>\%
  code_calls("head") \%>\%
  code_calls("mpg", "oops, wrong data set.")

# If in a tutor document
.answer \%>\%
  code_calls("head") \%>\%
  code_calls("mpg", "oops, wrong data set.")
}
}
