% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_message.R
\name{extract_message}
\alias{extract_message}
\title{Extract message}
\usage{
extract_message(result)
}
\arguments{
\item{result}{A result class object created by a magrader function. At the
moment the only result class is "failure".}
}
\value{
A character string that contains the message to display to the
  student.
}
\description{
Extracts the message component of a magrader pipe result
}
