% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_with.R
\name{fail_with}
\alias{fail_with}
\title{Create a failure message}
\usage{
fail_with(message)
}
\arguments{
\item{message}{A character string to display to the student.}
}
\value{
A character string with the S3 class "failure"
}
\description{
Returns a feedback message that can be piped through downstream checks
without triggering the checks.
}
\examples{
fail_with("Did you remember to use na.rm = TRUE?")
}
