% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Extracts code from a magrader submission object}
\usage{
get_code(submission)
}
\arguments{
\item{submission}{A character string or magrader submission object.}
}
\value{
A character string of student code supplied by the tutor package
}
\description{
Extracts code from a magrader submission object
}
