% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_failure.R
\name{is_failure}
\alias{is_failure}
\title{Does an object inherit the failure class?}
\usage{
is_failure(ans)
}
\arguments{
\item{ans}{A character string of student code provided by tutor or reults
class object created by a magrader function.}
}
\value{
`TRUE` or `FALSE`
}
\description{
`is_failure` can be used to check whether an answer has already been marked
as a failure by an upstream check in a check pipe.
}
