% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_submission.R
\name{is_submission}
\alias{is_submission}
\title{Is x a submission}
\usage{
is_submission(x)
}
\arguments{
\item{x}{an R object}
}
\value{
`TRUE` or `FALSE`
}
\description{
Tests whether an object inherits the magrader submission class. Submission
objects contain student code supplied by the tutor package as a character
string, a status, and possible feedback messages.
}
