% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_with.R
\name{pass_with}
\alias{pass_with}
\title{Mark a submission as a pass}
\usage{
pass_with(submission, praise = NULL)
}
\arguments{
\item{submission}{Student code provided as a character string by tutor, or a
submission object created by magrader}

\item{praise}{A character string to display to the student}
}
\value{
An S3 submission object
}
\description{
Returns a magrader submission object that can be piped to downstream checks
for additional checking.
}
