% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_checker.R
\name{pipe_checker}
\alias{pipe_checker}
\title{magrader interface to tutor}
\usage{
pipe_checker(label = NULL, user_code = NULL, check_code = NULL,
  envir_result = NULL, evaluate_result = NULL, ...)
}
\arguments{
\item{label}{The label of the code chunk to check}

\item{user_code}{The code provided by the student in the exercise, as a
character string}

\item{check_code}{The code supplied by the author in the `-check` or
`-code-check` chunks, as a character string}

\item{envir_result}{The R environment after executing the chunk.}

\item{evaluate_result}{The return value from the `evaluate::evaluate`
function.}

\item{...}{Unused (include for compatibility with parameters to be added in
the future)}
}
\value{
A list in the format that tutor requires to display student feedback.
}
\description{
An exercise checker function to use with the R Markdown tutor package. This
function does nothing but check a chunk. It is designed to be used by tutor;
tutor will call the function when necessary and supply all of the inputs. To
enable the function with tutor, set the following option in a code chunk,
`tutor_options(exercise.checker = checker)`.
}
