% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_with.R
\name{warn_with}
\alias{warn_with}
\title{Mark a submission as a pass with a warning}
\usage{
warn_with(submission, message = NULL)
}
\arguments{
\item{submission}{Student code provided as a character string by tutor, or a
submission object created by magrader}

\item{message}{A character string to display to the student}
}
\value{
An S3 submission object
}
\description{
Returns a magrader submission object that can be piped to downstream checks
for additional checking. If the submission does not fail any downstream
checks, it will be marked as a pass, but displayed with a warning message.
}
