% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_with.R
\name{fail_with}
\alias{fail_with}
\title{Mark a submission as a failure}
\usage{
fail_with(submission, message = NULL)
}
\arguments{
\item{submission}{Student code provided as a character string by tutor, or a
submission object created by magrader}

\item{message}{A character string to display to the student}
}
\value{
A magrader submission object with the status "fail"
}
\description{
Returns a magrader submission object that can be piped through downstream checks
without triggering the checks.
}
